Refactored step type classes to use a flow_step base class and a connector_step base class.
