+++ classes/step/flow_step.php
@@ -18,18 +18,38 @@ namespace tool_dataflows\step;
 
 use tool_dataflows\execution\engine;
 use tool_dataflows\execution\engine_step;
-use tool_dataflows\execution\engine_flow_cap;
+use tool_dataflows\execution\iterators\iterator;
+use tool_dataflows\execution\iterators\map_iterator;
+use tool_dataflows\execution\flow_engine_step;
 
 /**
- * A special, virtual flow step that is attached to the end of a flow block.
+ * Base class for flow step types.
  *
  * @package   tool_dataflows
  * @author    Jason den Dulk <jasondendulk@catalyst-au.net>
  * @copyright 2022, Catalyst IT
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+abstract class flow_step extends base_step {
 
-class flow_cap extends flow_step {
+    /**
+     * Does this type define a flow step?
+     * @return bool
+     */
+    final public function is_flow(): bool {
+        return true;
+    }
+
+    /**
+     * Step callback handler
+     *
+     * Implementation can vary, this might be a transformer, resource, or
+     * something else.
+     */
+    public function execute($input) {
+        // Default is to do nothing.
+        return $input;
+    }
 
     /**
      * Generates an engine step for this type.
@@ -39,6 +59,19 @@ class flow_cap extends flow_step {
      * @return engine_step
      */
     public function get_engine_step(engine $engine, \tool_dataflows\step $stepdef): engine_step {
-        return new engine_flow_cap($engine, $stepdef, $this);
+        // This should be sufficient for most cases. Override this function if needed.
+        return new flow_engine_step($engine, $stepdef, $this);
+    }
+
+    /**
+     * Get the iterator for the step, based on configurations.
+     *
+     * @param flow_engine_step $step
+     * @return iterator
+     */
+    public function get_iterator(flow_engine_step $step): iterator {
+        // Default is to simply map.
+        $input = current($step->upstreams)->iterator;
+        return new map_iterator($step, $input);
     }
 }
