+++ classes/step/void_step.php
@@ -28,15 +28,7 @@ use tool_dataflows\execution\flow_engine_step;
  * @copyright  Catalyst IT, 2022
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class void_step extends base_step {
-
-    /**
-     * Does this type define a flow step?
-     * @return bool
-     */
-    public function is_flow(): bool {
-        return true;
-    }
+class void_step extends flow_step {
 
     /**
      * Executes the step
@@ -48,17 +40,5 @@ class void_step extends base_step {
      */
     public function execute($input) {
         return;
-    }
-
-    /**
-     * Generates an engine step for this type.
-     *
-     * @param engine $engine
-     * @param \tool_dataflows\step $stepdef
-     * @return engine_step
-     */
-    public function get_engine_step(engine $engine, \tool_dataflows\step $stepdef): engine_step {
-        // This should be sufficient for most cases. Override this function if needed.
-        return new flow_engine_step($engine, $stepdef, $this);
     }
 }
